Moved blog page JS into the page, since it's specific to that page. Removed H1s from error pages
