+++ public_html/errors/403.php
@@ -4,6 +4,4 @@ $title = '403 Forbidden';
     
 ?>
 
-<h1>403 Forbidden</h1>
-
 <p>Directory access forbidden.</p>
