+++ public_html/errors/404.php
@@ -4,7 +4,5 @@ $title = '404 Not Found';
 
 ?>
 
-<h1>404 Not Found</h1>
-
 <p>Page does not exist.</p>
 <p>You may have typed an invalid URL, or the page has been removed.</p>
